requirements.txt
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.7.4
pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
flask>=2.3.0
flask-cors>=4.0.0
scikit-learn>=1.3.0